fix(Shopping): atomization of the shopping center's storage services
