+++ Poliedro.Eds.Api/Program.cs
@@ -217,7 +217,6 @@ builder.Services.AddScoped<IGetHoseNumber, GetIdAuxService>();
 builder.Services.AddScoped<IGetDispenserNumber, GetIdAuxService>();
 
 builder.Services.AddScoped<IProductCompartimentGetByCompartmentId, ProductCompartimentGetByCompartmentId>();
-builder.Services.AddScoped<IProductCompartimentStockUpdate, ProductCompartimentStockUpdateService>();
 
 builder.Services.AddScoped<ICourtListDomainService, CourtListService>();
 builder.Services.AddScoped<IInventoryListDomainService, InventoryListService>();
