+++ Poliedro.Eds.Application/Shopping/Commands/CreateShopping/CreateShoppingCommandHandler.cs
@@ -9,19 +9,15 @@ using Poliedro.Eds.Domain.Common.Results;
 using Poliedro.Eds.Domain.Common.Results.Errors;
 using Poliedro.Eds.Domain.Inventory.Entities;
 using Poliedro.Eds.Domain.Product.Entities;
-using Poliedro.Eds.Domain.ProductCompartiment.DomainProductCompartiment;
 using Poliedro.Eds.Domain.Shopping.DomainShopping;
 using Poliedro.Eds.Domain.Shopping.Entities;
 
 namespace Poliedro.Eds.Application.Shopping.Commands.CreateShopping;
 
 public class CreateShoppingCommandHandler(
-    IShoppingCreateShopping shoppingDomainService,
+    IShoppingTransactionalService shoppingTransactionalService,
     IMapper mapper,
-    IValidator<CreateShoppingRequestDto> validator,
-    IProductPriceUpdateService productPriceUpdateService,
-    IProductCompartimentStockUpdate productCompartimentStockUpdateService,
-    IRedisService redisService
+    IValidator<CreateShoppingRequestDto> validator
 ) : IRequestHandler<CreateShoppingCommand, Result<VoidResult, Error>>
 {
     public async Task<Result<VoidResult, Error>> Handle(CreateShoppingCommand request, CancellationToken cancellationToken)
@@ -33,33 +29,23 @@ public class CreateShoppingCommandHandler(
 
         var shoppingEntity = mapper.Map<ShoppingEntity>(request.Request);
 
-
-        if (request.Request.SellPriceProducts is { } sellPriceProducts && sellPriceProducts.Any())
-        {
-            var products = mapper.Map<IEnumerable<ProductEntity>>(sellPriceProducts);
-            var priceUpdateResult = await productPriceUpdateService.UpdatePricesAsync(products);
-            if (!priceUpdateResult.IsSuccess)
-                return priceUpdateResult.Error!;
-        }
-
-        if (shoppingEntity.ShoppingProducts is { } shoppingProducts && shoppingProducts.Any())
-        {
-            var stockUpdateResult = await productCompartimentStockUpdateService.UpdateStockAsync(shoppingProducts);
-            if (!stockUpdateResult.IsSuccess)
-                return stockUpdateResult.Error!;
-        }
-
         shoppingEntity.ShoppingInventory = new InventoryEntity
         {
             Date = DateOnly.FromDateTime(shoppingEntity.Date),
             ReferenceType = "shopping",
         };
 
-        var result = await shoppingDomainService.CreateAsync(shoppingEntity);
+        var productsToUpdatePrice = request.Request.ShoppingProducts
+            .Where(sp => sp.SellPrice.HasValue)
+            .Select(sp => new ProductEntity
+            {
+                IdProduct = sp.IdProduct,
+                Price = sp.SellPrice.Value
+            })
+            .ToList();
 
-        await RedisHelper.RemoveCacheIfSuccessAsync(result, redisService,KeyRedisConstants.SHOPPING);
-        await RedisHelper.RemoveCacheIfSuccessAsync(result, redisService, KeyRedisConstants.PRODUCT);
-
-        return result.IsSuccess ? result.Value! : result.Error!;
+        return await shoppingTransactionalService.ExecuteShoppingTransactionAsync(
+            shoppingEntity,
+            productsToUpdatePrice);
     }
 }
